move less essential binning functions back to peanut
